error en el gnl

+++ include/Libft/strings/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: agrimald <agrimald@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/18 15:09:07 by agrimald          #+#    #+#             */
-/*   Updated: 2023/09/15 20:19:41 by agrimald         ###   ########.fr       */
+/*   Updated: 2023/09/18 21:28:14 by agrimald         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
